+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/infrastructure/config/GovNotifyConfiguration.java
@@ -7,9 +7,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.notify.CustomNotificationClient;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.notify.NotificationClientApi;
-import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.notify.RetryableNotificationClient;
+import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.clients.RetryableNotificationClient;
+import uk.gov.service.notify.CustomNotificationClient;
 
 @Slf4j
 @Configuration
@@ -17,7 +16,7 @@ public class GovNotifyConfiguration {
 
     @Bean
     @Primary
-    public NotificationClientApi notificationClient(
+    public RetryableNotificationClient notificationClient(
         @Value("${govnotify.key}") String key,
         @Value("${govnotify.baseUrl}") String goveNotifyBaseUrl,
         @Value("${govnotify.timeout}") int timeout
